start of 1.21.6 compatibility
The test for FlyingMob was removed since I could not figure out what
it should be in 1.21.6. This needs to be fixed properly. Other than that
this is mostly a version bump and fixing some function signatures.

This version seems to fulfil our needs, but only rules we actually use
on any server were tested.